+++ js/tokens.ts
@@ -79,9 +79,7 @@ export type TUpdateKind = void | number;
 export interface IBehave extends WidgetModel {
   rank: number;
   updateRequested: ISignal<IBehave, TUpdateKind>;
-  graphDataUpdateRequested: ISignal<IBehave, void>;
   extraColumns?: IExtraColumns;
-  updateGraphData?(graphData: GraphData): Promise<void>;
   // link
   getLinkColor?(options: ILinkBehaveOptions): string | null;
   getLinkCurvature?(options: ILinkBehaveOptions): number | null;
@@ -257,6 +255,8 @@ export enum ECoerce {
   boolish = 'boolean',
   numeric = 'number',
 }
+
+export type TCoercer = (value: any) => any;
 
 export interface IExtraColumns {
   nodes: string[];